requirements.txt
crawl4ai
pydantic
python-dotenv
aiohttp
cryptography
